Moved Toolbar and Upgrade to Pro
